Split storage_access_record
Split storage_access_record into storage_read_record and storage_write_record
