+++ src/trace/mod.rs
@@ -1,5 +1,6 @@
 pub mod state_trace_tuple;
-pub mod storage_access_record;
 pub mod state_trace_manager;
 pub mod section_type;
 pub mod storage_type;
+pub mod storage_read_record;
+pub mod storage_write_record;
