+++ src/trace/state_trace_tuple.rs
@@ -1,6 +1,7 @@
-use crate::trace::storage_access_record::{
-    StorageReadRecord, StorageWriteRecord
+use crate::trace::storage_read_record::{
+    StorageReadRecord
 };
+use crate::trace::storage_write_record::StorageWriteRecord;
 
 pub const MAX_NUM_READ_LOCATIONS: usize = 9;
 pub const MAX_NUM_WRITE_LOCATIONS: usize = 2;
