+++ src/trace/storage_write_record.rs
@@ -1,0 +1,31 @@
+use crate::trace::section_type::SectionType;
+use crate::trace::storage_type::StorageType;
+
+#[derive(Debug)]
+pub struct StorageWriteRecord {
+    storage_type: StorageType,
+    section_type: SectionType,
+    location: u64,
+    value: u64,
+    time_stamp: u64,
+    // access_type: AccessType,
+}
+
+impl StorageWriteRecord {
+    pub fn new(storage_type: StorageType,
+               section_type: SectionType,
+               location: u64,
+               value: u64,
+               time_stamp: u64,
+               // access_type: AccessType
+    ) -> Self {
+        Self {
+            storage_type,
+            section_type,
+            location,
+            value,
+            time_stamp,
+            // access_type: access_type,
+        }
+    }
+}
